# food/src/main/java/hu/oe/yokudlela/food/generated/rest/api/DefaultApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package hu.oe.yokudlela.food.generated.rest.api;

import hu.oe.yokudlela.food.generated.rest.model.ErrorResponse;
import hu.oe.yokudlela.food.generated.rest.model.FoodCategoryRequest;
import hu.oe.yokudlela.food.generated.rest.model.FoodCategoryResponse;
import hu.oe.yokudlela.food.generated.rest.model.FoodRequest;
import hu.oe.yokudlela.food.generated.rest.model.FoodResponse;
import hu.oe.yokudlela.food.generated.rest.model.IdModel;

import jakarta.annotation.security.RolesAllowed;
import org.springframework.security.access.prepost.PreAuthorize;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
@Tag(name = "Default", description = "the Default API")
public interface DefaultApi {

    /**

     * GET /categories : Kategóriák listázása
     * Lekérdezi az összes ételkategóriát.
     *
     * @return Kategóriák sikeresen lekérdezve (status code 200)
     */




    @Operation(
        operationId = "categoriesGet",
        summary = "Kategóriák listázása",
        description = "Lekérdezi az összes ételkategóriát.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Kategóriák sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = FoodCategoryResponse.class)))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/categories",
        produces = { "application/json" }
    )
    @RolesAllowed("view")



    
    ResponseEntity<List<FoodCategoryResponse>> categoriesGet(
    );

    /**

     * DELETE /categories/{id} : Kategória törlése
     * Törli a kategóriát ID alapján.
     *
     * @param id  (required)
     * @return Kategória törölve (status code 204)
     *         or Kategória nem található (status code 404)
     */




    @Operation(
        operationId = "categoriesIdDelete",
        summary = "Kategória törlése",
        description = "Törli a kategóriát ID alapján.",
        responses = {
            @ApiResponse(responseCode = "204", description = "Kategória törölve"),
            @ApiResponse(responseCode = "404", description = "Kategória nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/categories/{id}",
        produces = { "application/json" }
    )
    @RolesAllowed("edit")



    
    ResponseEntity<Void> categoriesIdDelete(
        

@Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id 
    );

    /**

     * GET /categories/{id} : Kategória lekérdezése
     * Lekérdezi egy adott kategóriát ID alapján.
     *
     * @param id A kategória egyedi azonosítója. (required)
     * @return Kategória sikeresen lekérdezve (status code 200)
     *         or Kategória nem található (status code 404)
     */




    @Operation(
        operationId = "categoriesIdGet",
        summary = "Kategória lekérdezése",
        description = "Lekérdezi egy adott kategóriát ID alapján.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Kategória sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FoodCategoryResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Kategória nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/categories/{id}",
        produces = { "application/json" }
    )
    @RolesAllowed("view")



    
    ResponseEntity<FoodCategoryResponse> categoriesIdGet(
        

@Parameter(name = "id", description = "A kategória egyedi azonosítója.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id 
    );

    /**

     * POST /categories : Új kategória hozzáadása
     * Új ételkategóriát hoz létre.
     *
     * @param foodCategoryRequest  (required)
     * @return Kategória sikeresen létrehozva (status code 201)
     */




    @Operation(
        operationId = "categoriesPost",
        summary = "Új kategória hozzáadása",
        description = "Új ételkategóriát hoz létre.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Kategória sikeresen létrehozva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = IdModel.class))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/categories",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    @RolesAllowed("edit")



    
    ResponseEntity<IdModel> categoriesPost(
        

 @Parameter(name = "FoodCategoryRequest", description = "", required = true) @Valid @RequestBody FoodCategoryRequest foodCategoryRequest
    );

    /**

     * GET /foods : Ételek listázása
     * Lekérdezi az összes ételt.
     *
     * @return Ételek sikeresen lekérdezve (status code 200)
     */




    @Operation(
        operationId = "foodsGet",
        summary = "Ételek listázása",
        description = "Lekérdezi az összes ételt.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Ételek sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = FoodResponse.class)))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/foods",
        produces = { "application/json" }
    )
    @RolesAllowed("view")



    
    ResponseEntity<List<FoodResponse>> foodsGet(
    );

    /**

     * DELETE /foods/{id} : Étel törlése
     * Törli az ételt az ID alapján.
     *
     * @param id  (required)
     * @return Étel törölve (status code 204)
     *         or Étel nem található (status code 404)
     */




    @Operation(
        operationId = "foodsIdDelete",
        summary = "Étel törlése",
        description = "Törli az ételt az ID alapján.",
        responses = {
            @ApiResponse(responseCode = "204", description = "Étel törölve"),
            @ApiResponse(responseCode = "404", description = "Étel nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/foods/{id}",
        produces = { "application/json" }
    )
    @RolesAllowed("edit")



    
    ResponseEntity<Void> foodsIdDelete(
        

@Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id 
    );

    /**

     * GET /foods/{id} : Étel lekérdezése
     * Lekérdez egy ételt az ID alapján.
     *
     * @param id  (required)
     * @return Étel sikeresen lekérdezve (status code 200)
     *         or Étel nem található (status code 404)
     */




    @Operation(
        operationId = "foodsIdGet",
        summary = "Étel lekérdezése",
        description = "Lekérdez egy ételt az ID alapján.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Étel sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FoodResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Étel nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/foods/{id}",
        produces = { "application/json" }
    )
    @RolesAllowed("view")



    
    ResponseEntity<FoodResponse> foodsIdGet(
        

@Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id 
    );

    /**

     * PUT /foods/{id} : Étel módosítása
     * Módosítja az étel adatait.
     *
     * @param id  (required)
     * @param foodRequest  (required)
     * @return Étel sikeresen módosítva (status code 200)
     *         or Étel nem található (status code 404)
     */




    @Operation(
        operationId = "foodsIdPut",
        summary = "Étel módosítása",
        description = "Módosítja az étel adatait.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Étel sikeresen módosítva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FoodResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Étel nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/foods/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    @RolesAllowed("edit")



    
    ResponseEntity<FoodResponse> foodsIdPut(
        

@Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id 
        ,

 @Parameter(name = "FoodRequest", description = "", required = true) @Valid @RequestBody FoodRequest foodRequest
    );

    /**

     * POST /foods : Új étel hozzáadása
     * Új ételt ad hozzá a rendszerhez.
     *
     * @param foodRequest  (required)
     * @return Étel sikeresen hozzáadva (status code 201)
     */




    @Operation(
        operationId = "foodsPost",
        summary = "Új étel hozzáadása",
        description = "Új ételt ad hozzá a rendszerhez.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Étel sikeresen hozzáadva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = IdModel.class))
            })
        }
        ,
        security = {
            @SecurityRequirement(name = "OpenAPI", scopes={ })
        }

        
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/foods",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    @RolesAllowed("edit")



    
    ResponseEntity<IdModel> foodsPost(
        

 @Parameter(name = "FoodRequest", description = "", required = true) @Valid @RequestBody FoodRequest foodRequest
    );

}