# food/src/main/java/hu/oe/yokudlela/food/generated/rest/model/FoodCategoryRequest.java
package hu.oe.yokudlela.food.generated.rest.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import org.openapitools.jackson.nullable.JsonNullable;
import java.io.Serializable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import org.hibernate.validator.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

/**
* FoodCategoryRequest
*/

@lombok.Data
@lombok.NoArgsConstructor
@lombok.AllArgsConstructor
@lombok.Builder

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
public class FoodCategoryRequest implements Serializable {

    private static final long serialVersionUID = 1L;

        @hu.oe.yokudlela.food.validation.NameExists(message="error.food.name.exists")


            private String name;
    public static final String FIELD_NAME="name";

    }