# food/src/main/resources/openapi/food.yaml
openapi: 3.0.0
info:
  title: Étel Rendelési API
  description: "API ételek és ételkategóriák nyilvántartására és kezelésére."
  version: 1.0.0

servers:
  - url: https://example-food-api.cloud/api
  - url: http://localhost:8080/api
    description: Helyi fejlesztői szerver

paths:
  /categories:
    get:
      summary: Kategóriák listázása
      description: Lekérdezi az összes ételkategóriát.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("view")'
      responses:
        "200":
          description: Kategóriák sikeresen lekérdezve
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodCategoryResponse"
              example:
                - id: "1"
                  name: Levesek
                - id: "2"
                  name: Főételek

    post:
      summary: Új kategória hozzáadása
      description: Új ételkategóriát hoz létre.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("edit")'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodCategoryRequest"
      responses:
        "201":
          description: Kategória sikeresen létrehozva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdModel"
              example:
                id: "4"

  /categories/{id}:
    get:
      summary: Kategória lekérdezése
      description: Lekérdezi egy adott kategóriát ID alapján.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("view")'
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: A kategória egyedi azonosítója.
      responses:
        "200":
          description: Kategória sikeresen lekérdezve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodCategoryResponse"
              example:
                id: "1"
                name: Levesek
        "404":
          description: Kategória nem található
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Kategória törlése
      description: Törli a kategóriát ID alapján.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("edit")'
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Kategória törölve
        "404":
          description: Kategória nem található
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /foods:
    get:
      summary: Ételek listázása
      description: Lekérdezi az összes ételt.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("view")'
      responses:
        "200":
          description: Ételek sikeresen lekérdezve
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodResponse"
              example:
                - id: "1"
                  name: Gulyásleves
                  categoryId: "1"
                  price: 1200

    post:
      summary: Új étel hozzáadása
      description: Új ételt ad hozzá a rendszerhez.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("edit")'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodRequest"
      responses:
        "201":
          description: Étel sikeresen hozzáadva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdModel"
              example:
                id: "4"

  /foods/{id}:
    get:
      summary: Étel lekérdezése
      description: Lekérdez egy ételt az ID alapján.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("view")'
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Étel sikeresen lekérdezve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodResponse"
        "404":
          description: Étel nem található
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Étel módosítása
      description: Módosítja az étel adatait.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("edit")'
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodRequest"
      responses:
        "200":
          description: Étel sikeresen módosítva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodResponse"
              example:
                id: "1"
                name: Húsleves
                categoryId: "1"
                price: 1300
        "404":
          description: Étel nem található
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Étel törlése
      description: Törli az ételt az ID alapján.
      security:
        - BearerAuth: [ ]
      x-operation-extra-annotation: '@RolesAllowed("edit")'
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Étel törölve
        "404":
          description: Étel nem található
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    IdModel:
      type: object
      properties:
        id:
          type: string
          description: Egyedi azonosító.

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Erőforrás nem található

    FoodCategory:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-field-extra-annotation: "@jakarta.persistence.Id @jakarta.persistence.GeneratedValue(strategy = jakarta.persistence.GenerationType.IDENTITY)"
        name:
          type: string
          x-field-extra-annotation: '@hu.oe.yokudlela.food.validation.NameExists(message="error.food.name.exists")'

    FoodCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          x-field-extra-annotation: '@hu.oe.yokudlela.food.validation.NameExists(message="error.food.name.exists")'

    FoodCategoryResponse:
      allOf:
        - $ref: '#/components/schemas/IdModel'
        - $ref: '#/components/schemas/FoodCategoryRequest'

    Food:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-field-extra-annotation: "@jakarta.persistence.Id @jakarta.persistence.GeneratedValue(strategy = jakarta.persistence.GenerationType.UUID)"
        name:
          type: string
        categoryId:
          type: string
        price:
          type: integer

    FoodRequest:
      type: object
      required:
        - name
        - categoryId
        - price
      properties:
        name:
          type: string
          x-field-extra-annotation: '@hu.oe.yokudlela.food.validation.NameExists(message="error.food.name.exists")'
        categoryId:
          type: string
        price:
          type: integer

    FoodResponse:
      allOf:
        - $ref: '#/components/schemas/IdModel'
        - $ref: '#/components/schemas/FoodRequest'

security:
  - BearerAuth: [ ]